feat(LogsVolumePanel): decouple filter logic from scene
